remove CouchbaseRequestInfoHolder
