--- instrumentation/couchbase/couchbase-2-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/couchbase/v2_0/CouchbaseRequestInfoHolder.java
@@ -1,29 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.couchbase.v2_0;
-
-import static io.opentelemetry.context.ContextKey.named;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.ContextKey;
-import javax.annotation.Nullable;
-
-public final class CouchbaseRequestInfoHolder {
-
-  private static final ContextKey<CouchbaseRequestInfo> KEY =
-      named("opentelemetry-couchbase-request-key");
-
-  private CouchbaseRequestInfoHolder() {}
-
-  public static Context init(Context context, CouchbaseRequestInfo couchbaseRequest) {
-    return context.with(KEY, couchbaseRequest);
-  }
-
-  @Nullable
-  public static CouchbaseRequestInfo get(Context context) {
-    return context.get(KEY);
-  }
-}
